Minor bugfix and added bookdown files

+++ R/sync.R
@@ -30,6 +30,8 @@ sync <- function(new, old) {
   additions <- 0L
   deletes <- 0L
 
+  rlang::inform("The following changes will be made")
+
   # go through every file in the arrivals area
   for (i in seq_along(arrivals)) {
     new_file <- paste0(new, arrivals[i])
@@ -39,14 +41,10 @@ sync <- function(new, old) {
       # check if the files differ. We can use Hashes for this basic functionality
       if (as.character(tools::md5sum(new_file)) != as.character(tools::md5sum(old_file))) {
         rlang::inform(glue::glue("changes detected: overwriting {new_file}"))
-        invisible(file.copy(from = new_file, to = old_file, overwrite = TRUE))
-        overwrites <- overwrites + 1L
       }
       # Otherwise if file is not in existing (i.e. it is a new file) copy straight across
     } else {
       rlang::inform(glue::glue("new file: copying {new_file}"))
-      invisible(file.copy(from = new_file, to = paste0(old, arrivals[i])))
-      additions <- additions + 1L
     }
   }
 
@@ -55,15 +53,48 @@ sync <- function(new, old) {
     existing_file <- paste0(old, existing[j])
     if (!(existing[j] %in% arrivals)) {
       rlang::inform(glue::glue("file removed: deleting {existing_file}"))
-      invisible(file.remove(existing_file))
-      deletes <- deletes + 1L
     }
   }
 
-  rlang::inform(glue::glue(
-    "summary
+  x <- readline(prompt = "Are you happy to proceed? 1 = Yes, Any = No: ")
+
+  if (x == "1") {
+    # go through every file in the arrivals area
+    for (i in seq_along(arrivals)) {
+      new_file <- paste0(new, arrivals[i])
+      # check if arrival file already exists
+      if (arrivals[i] %in% existing) {
+        old_file <- paste0(old, arrivals[i])
+        # check if the files differ. We can use Hashes for this basic functionality
+        if (as.character(tools::md5sum(new_file)) != as.character(tools::md5sum(old_file))) {
+          rlang::inform(glue::glue("changes detected: overwriting {new_file}"))
+          invisible(file.copy(from = new_file, to = old_file, overwrite = TRUE))
+          overwrites <- overwrites + 1L
+        }
+        # Otherwise if file is not in existing (i.e. it is a new file) copy straight across
+      } else {
+        rlang::inform(glue::glue("new file: copying {new_file}"))
+        invisible(file.copy(from = new_file, to = paste0(old, arrivals[i])))
+        additions <- additions + 1L
+      }
+    }
+
+    # Now clean up any files that have been deleted
+    for (j in seq_along(existing)) {
+      existing_file <- paste0(old, existing[j])
+      if (!(existing[j] %in% arrivals)) {
+        rlang::inform(glue::glue("file removed: deleting {existing_file}"))
+        invisible(file.remove(existing_file))
+        deletes <- deletes + 1L
+      }
+    }
+
+    rlang::inform(glue::glue(
+      "summary
     additions: {additions}
     deletions: {deletes}
     overwrites: {overwrites}"))
-
+  } else {
+    rlang::inform("Aborting")
+  }
 }
